+++ fiddles/react/fiddle-0001-Element/src/index.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import './index.css';
+
+function tick() {
+  const element = (
+    <div>
+      <h1>{new Date().toLocaleTimeString()}</h1>
+    </div>
+  );
+  ReactDOM.render(element, document.getElementById('root'));
+}
+
+setInterval(tick, 1000);