+++ fiddles/react/fiddle-0005-EventHandler/src/toggle/Toggle.js
@@ -1,0 +1,28 @@
+import React, { Component } from 'react';
+import './Toggle.css';
+
+class Toggle extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {isToggleOn: true};
+    // This binding is necessary to make `this` work in the callback
+    this.handleClick = this.handleClick.bind(this);
+  }
+
+  handleClick() {
+    this.setState(prevState => ({
+      isToggleOn: !prevState.isToggleOn
+    }));
+  }
+
+  render() {
+    return (
+      <button className={this.state.isToggleOn ? 'toggle green' : 'toggle red'} onClick={this.handleClick}>
+        {this.state.isToggleOn ? 'ON' : 'OFF'}
+      </button>
+    );
+  }
+}
+
+
+export default Toggle;