+++ fiddles/react/fiddle-0007-HiddenRendering/src/App.js
@@ -1,0 +1,16 @@
+import React, { Component } from 'react';
+import './App.css';
+import Page from './page/Page';
+
+
+class App extends Component {
+  render() {
+    return (
+      <div className="App">
+        <Page/>
+      </div>
+    );
+  }
+}
+
+export default App;