+++ fiddles/react/fiddle-0008-ListComponent/src/blog/Blog.test.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import Blog from './Blog';
+
+it('renders without crashing', () => {
+  const div = document.createElement('div');
+  ReactDOM.render(<Blog />, div);
+  ReactDOM.unmountComponentAtNode(div);
+});