+++ fiddles/react/fiddle-0009-Form/src/App.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+import './App.css';
+import NameForm from './name-form/NameForm';
+
+class App extends Component {
+  render() {
+    return (
+      <NameForm/>
+    );
+  }
+}
+
+export default App;