+++ fiddles/react/fiddle-0010-TextAreaControl/src/App.js
@@ -1,0 +1,16 @@
+import React, { Component } from 'react';
+import './App.css';
+import './essay-form/EssayForm';
+import EssayForm from "./essay-form/EssayForm";
+
+class App extends Component {
+  render() {
+    return (
+      <div className="App">
+        <EssayForm/>
+      </div>
+    );
+  }
+}
+
+export default App;