+++ fiddles/react/fiddle-0012-MultipleInputs/src/App.js
@@ -1,0 +1,15 @@
+import React, { Component } from 'react';
+import './App.css';
+import ReservationForm from './reservation-form/ReservationForm';
+
+class App extends Component {
+  render() {
+    return (
+      <div className="App">
+        <ReservationForm/>
+      </div>
+    );
+  }
+}
+
+export default App;