+++ fiddles/react/fiddle-0013-StateLifting/src/App.js
@@ -1,0 +1,15 @@
+import React, { Component } from 'react';
+import './App.css';
+import Calculator from './calculator/Calculator';
+
+class App extends Component {
+  render() {
+    return (
+      <div className="App">
+        <Calculator/>
+      </div>
+    );
+  }
+}
+
+export default App;